+++ lib/src/analytics/analytics_interface.dart
@@ -28,7 +28,8 @@ abstract class AnalyticsInterface {
 }
 
 class IndividualUsage {
-  late final Loglytics loglytics = Loglytics.create(location: runtimeType.toString());
+  late final Loglytics loglytics =
+      Loglytics.create(location: runtimeType.toString());
   late final Log log = Log(location: runtimeType.toString());
   late final AnalyticsService analyticsService = AnalyticsService();
 }